refactor(llm-agent): move IClientAdapter and ILlmApiAdapter from llm-agent-server
- Move api-adapter.ts (AdapterValidationError, ApiRequestContext, ApiSseEvent, ILlmApiAdapter, NormalizedRequest) to packages/llm-agent/src/interfaces/
- Move client-adapter.ts (IClientAdapter) to packages/llm-agent/src/interfaces/
- Replace server copies with re-export shims from @mcp-abap-adt/llm-agent
